Refactor plugin utils.
