+++ kolibri/core/models.py
@@ -4,8 +4,8 @@ from __future__ import unicode_literals
 
 from django.conf import settings
 
-from kolibri.plugins.registry import register
+from kolibri.plugins.registry import registered_plugins
 
 # Register any django apps that may have kolibri plugin
 # modules inside them
-register(settings.INSTALLED_APPS)
+registered_plugins.register(settings.INSTALLED_APPS)
